langchain-rag-alice-chatbot/backend/backend.py
# backend.py
from dotenv import load_dotenv
import os

# Load environment variables from .env file
load_dotenv()

from fastapi import FastAPI, Request
from pydantic import BaseModel
from langchain_chroma import Chroma
from langchain_openai import OpenAIEmbeddings, ChatOpenAI
from langchain.prompts import ChatPromptTemplate
from fastapi.middleware.cors import CORSMiddleware

app = FastAPI()

# Configure CORS
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],  # Allows all origins
    allow_credentials=True,
    allow_methods=["*"],  # Allows all methods
    allow_headers=["*"],  # Allows all headers
)

class Question(BaseModel):
    query: str

CHROMA_PATH = "chroma"
PROMPT_TEMPLATE = """
Answer the question using only the following context:

{context}

---

Question: {question}
"""

@app.post("/ask")
async def ask_question(data: Question):
    embedding_function = OpenAIEmbeddings()
    db = Chroma(persist_directory=CHROMA_PATH, embedding_function=embedding_function)
    results = db.similarity_search_with_relevance_scores(data.query, k=3)

    if not results or results[0][1] < 0.7:
        return {"answer": "No relevant information found."}

    context_text = "\n\n---\n\n".join([doc.page_content for doc, _ in results])
    prompt = ChatPromptTemplate.from_template(PROMPT_TEMPLATE).format(
        context=context_text, question=data.query
    )

    model = ChatOpenAI()
    response = model.invoke(prompt).content

    return {"answer": response}
