langchain-rag-alice-chatbot/backend/requirements.txt
python-dotenv
langchain
langchain-community
langchain-openai
langchain-chroma
openai
tiktoken

chromadb

unstructured
unstructured-client>=0.35.0
python-magic
lxml
nltk
beautifulsoup4
emoji
python-iso639
langdetect
rapidfuzz
backoff
psutil
html5lib
filetype
chardet

fastapi
uvicorn
